Fix the form of the sum and compute gamma
The sum computed tensor for the density of injected carriers, instead of the actual current density tenosr. Also width of the Lorentz function was set inside the code.

The dimension of the output tensor was changed from 3 * 3 * 3 to 3 * 3 * 3 * 3 and a few new terms were intruduced to the sum. Also a new function Get_gamma was added.

Gamma is computed as a frequency matching the distance of gamma point to its nearest neigbour in parabolica approximation.