Add argument parser
So far the arguments were parsed through the bash script. So the bash script and the actual computational python script were renamed and a new main script was added (will be renamed again).

Arguments are now parsed through argparse (8 in total). Frequency of the incident light is is mandatory for the run. The rest restricts number of bands and kpoints, decides the name of input/output and the number of processes for the parallel run.