Fix result dimension and a condition on kpoints
The programmme no longer test, wheather the band gap in given kpoint is smaller than 2omega, this condition
enforced singularities and non-continuity. The result also did not take into account the number of kpoints.

The condition on kpoints is completely removed. Furthermore, the result is now divided but the number of kpoints
(including weights).