Allow for writting the QI tensor to a file
The programme so far has only been able to print the result into the
stdout.
The main body of the programme was now moved to VASP2QI_filwriter.py,
which computes the QI tensor and writes it into a file. It check for
existing file, which are not overwritten and also can create the needed
directory.